Removido warning e traces do projeto
git-svn-id: https://mbsutil.googlecode.com/svn@109 6ba988c5-0224-0410-acf2-7d9395eead91

+++ Bufer.cpp
@@ -157,7 +157,7 @@ int Buffer::writeLong(unsigned long ul)
 
 int Buffer::readString(char *dest, int max)
 {
-	int tamstr=strlen((char*)pntL)+1;
+	unsigned long tamstr=strlen((char*)pntL)+1;
 	if(disponiveis()<tamstr)
 		return -1;
 	if(max<tamstr)
